+++ owlbot.py
@@ -25,13 +25,13 @@ for library in s.get_staging_dirs(default_version):
     # Avoid breaking change due to change in field renames.
     # https://github.com/googleapis/python-bigquery/issues/319
     s.replace(
-        library / f"google/cloud/bigquery_{library.name}/types/standard_sql.py",
+        library / f"google/cloud/bigquery_{library.name}/**/*.py",
         r"type_ ",
         "type ",
     )
     # Patch docs issue
     s.replace(
-        library / f"google/cloud/bigquery_{library.name}/types/model.py",
+        library / f"google/cloud/bigquery_{library.name}/**/*.py",
         r"""\"predicted_\"""",
         """`predicted_`""",
     )
